# [Core] Optimize get core worker Stats