--- a/src/ray/core_worker/store_provider/memory_store/memory_store.cc
+++ b/src/ray/core_worker/store_provider/memory_store/memory_store.cc
@@ -217,7 +217,8 @@ bool CoreWorkerMemoryStore::Put(const RayObject &object, const ObjectID &object_
       // If there is no existing get request, then add the `RayObject` to map.
       EmplaceObject(object_id, object_entry);
     } else {
-      OnErase(object_entry);
+      // It is equivalent to that object is added and immidiately deleted from the store.
+      OnDelete(object_entry);
     }
   }
 
@@ -283,7 +284,7 @@ Status CoreWorkerMemoryStore::GetImpl(const std::vector<ObjectID> &object_ids,
     // Clean up the objects if ref counting is off.
     if (ref_counter_ == nullptr) {
       for (const auto &object_id : ids_to_remove) {
-        objects_.erase(object_id);
+        EraseObject(object_id);
       }
     }
 
@@ -436,8 +437,8 @@ void CoreWorkerMemoryStore::Delete(const absl::flat_hash_set<ObjectID> &object_i
       if (it->second->IsInPlasmaError()) {
         plasma_ids_to_delete->insert(object_id);
       } else {
-        OnErase(it->second);
-        objects_.erase(it);
+        OnDelete(it->second);
+        EraseObject(object_id);
       }
     }
   }
@@ -448,8 +449,8 @@ void CoreWorkerMemoryStore::Delete(const std::vector<ObjectID> &object_ids) {
   for (const auto &object_id : object_ids) {
     auto it = objects_.find(object_id);
     if (it != objects_.end()) {
-      OnErase(it->second);
-      objects_.erase(it);
+      OnDelete(it->second);
+      EraseObject(object_id);
     }
   }
 }
@@ -466,7 +467,7 @@ bool CoreWorkerMemoryStore::Contains(const ObjectID &object_id, bool *in_plasma)
   return false;
 }
 
-void CoreWorkerMemoryStore::OnErase(std::shared_ptr<RayObject> obj) {
+void CoreWorkerMemoryStore::OnDelete(std::shared_ptr<RayObject> obj) {
   rpc::ErrorType error_type;
   // TODO(ekl) note that this doesn't warn on errors that are stored in plasma.
   if (obj->IsException(&error_type) &&
@@ -477,40 +478,46 @@ void CoreWorkerMemoryStore::OnErase(std::shared_ptr<RayObject> obj) {
       && !obj->WasAccessed() && unhandled_exception_handler_ != nullptr) {
     unhandled_exception_handler_(*obj);
   }
+}
 
-  if (obj->IsInPlasmaError()) {
+inline void CoreWorkerMemoryStore::EraseObject(const ObjectID &object_id) {
+  auto it = objects_.find(object_id);
+  if (it == objects_.end()) {
+    return;
+  }
+
+  objects_.erase(it);
+  if (it->second->IsInPlasmaError()) {
     num_in_plasma_ -= 1;
   } else {
     num_local_objects_ -= 1;
-    used_object_store_memory_ -= obj->GetSize();
+    used_object_store_memory_ -= it->second->GetSize();
   }
-  RAY_CHECK(num_in_plasma_ >= 0 && num_local_objects_ >= 0 && used_object_store_memory_ >= 0);
+  RAY_CHECK(num_in_plasma_ >= 0 && num_local_objects_ >= 0 &&
+            used_object_store_memory_ >= 0);
 }
 
-inline void CoreWorkerMemoryStore::EmplaceObject(const ObjectID &object_id, std::shared_ptr<RayObject> &object_entry) {
+inline void CoreWorkerMemoryStore::EmplaceObject(
+    const ObjectID &object_id, std::shared_ptr<RayObject> &object_entry) {
   auto inserted = objects_.emplace(object_id, object_entry).second;
   if (inserted) {
     if (object_entry->IsInPlasmaError()) {
       num_in_plasma_ += 1;
     } else {
-      num_local_objects_ += 1
+      num_local_objects_ += 1;
       used_object_store_memory_ += object_entry->GetSize();
     }
   }
-  RAY_CHECK(num_in_plasma_ >= 0 && num_local_objects_ >= 0 && used_object_store_memory_ >= 0);
+  RAY_CHECK(num_in_plasma_ >= 0 && num_local_objects_ >= 0 &&
+            used_object_store_memory_ >= 0);
 }
 
 MemoryStoreStats CoreWorkerMemoryStore::GetMemoryStoreStatisticalData() {
   absl::MutexLock lock(&mu_);
   MemoryStoreStats item;
-  for (const auto &it : objects_) {
-    if (it.second->IsInPlasmaError()) {
-      item.num_in_plasma += 1;
-    } else {
-      item.num_local_objects += 1;
-      item.used_object_store_memory += it.second->GetSize();
-    }
-  }
+  item.num_in_plasma = num_in_plasma_;
+  item.num_local_objects = num_local_objects_;
+  item.used_object_store_memory = used_object_store_memory_;
   return item;
 }