<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

Calculate the in memory store stats in flight instead of iterating all objects per second.

Before;
```
1620	CoreWorkerService.grpc_server.GetCoreWorkerStats - 50 total (0 active), CPU time: mean = 112.444 ms, total = 5.622 s
```
After
```
1793	CoreWorkerService.grpc_server.GetCoreWorkerStats - 62 total (0 active), CPU time: mean = 159.206 us, total = 9.871 ms
```

It is 1000X faster with the 1000 partition shuffle. 


## Related issue number

https://github.com/ray-project/ray/issues/14983

## Checks

- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(